+++ lib/quick_travel/client_templates.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'quick_travel/adapter'
-
-module QuickTravel
-  class ClientTemplates < Adapter
-    self.api_base = '/api/client_templates'
-  end
-end